refactor the code and fix initial bugs
